requirements.txt
pathspec
python-dotenv